implement result return back logic in endpoint (merge request !43)
Squash merge branch 'implement-mainloop5' into 'master'
1. implement result return back logic in endpoint

+++ src/numerous_llm/runtime/infer_request.h
@@ -70,7 +70,7 @@ class InferRequest {
 
  public:
   // The req id of the user's request.
-  int req_id;
+  int64_t req_id;
 
   // The unique id of this request.
   int infer_id;
